refactor: apply .editorconfig formatter

+++ src/main/java/com/peco2282/bcreborn/common/XorShift128Random.java
@@ -51,6 +51,6 @@ public class XorShift128Random {
   }
 
   public double nextDouble() {
-    return (double) (long) (nextLong() & 0x1FFFFFFFFFFFFFL) * DOUBLE_UNIT;
+    return (double) (nextLong() & 0x1FFFFFFFFFFFFFL) * DOUBLE_UNIT;
   }
 }
